chore: fix current megacheck errors
I spent some time trying to get megacheck into CI as well, but I ran
into continual problems, and I'm not able to spend any more time on it.

The current version of megacheck is not module aware. I tried a few
different approaches of copying the source to a new, traditional GOPATH
and running megacheck there, but something went wrong every time I did
it.

Then I found about about go-tools' `next` branch, which is module aware.
However, that version runs comment style checkers which aren't present
in megacheck, and I couldn't figure out how to disable that check.

So this commit only fixes the current megacheck errors without touching
the Circle config.

+++ semantic/types.go
@@ -271,7 +271,7 @@ func NewObjectType(propertyTypes map[string]Type) Type {
 
 		// track hash of property names and kinds
 		sum.Write([]byte(p))
-		binary.Write(sum, binary.LittleEndian, t.Nature())
+		binary.Write(sum, binary.LittleEndian, int64(t.Nature()))
 	}
 
 	// Create new object type
